requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.2
requests==2.32.3
python-dotenv==1.0.1
email-validator==2.2.0
werkzeug==3.1.3
gunicorn==23.0.0
Flask-Migrate==4.1.0
bleach==6.1.0